+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.4] 2021-11-18
+
+- Remove `OutputJson` because `Json` can replace it.
+- Allowed use validators on wrapper types, for example: `Option<T>`, `MaybeUnefined<T>`.
+
 ## [3.0.3] 2021-11-18
 
 - [integrations] Make `GraphQLWebSocket::new` use generic stream.